Matching to Old Repo Changes
Matching all the code changes from the old English translation repo that I am eventually making private since it is forked from the main PRET repo instead of my personal fork which will house many of the changes since I am pretty much the only one still doing anything with the demo ROM disassembly at this point. Here is a link to the old EN repo in case it is still up and you want to look at it for some reason.

https://github.com/WavelinkStudios/pokegold-spaceworld-en-old