Translations (New & Corrected) & ROMX Bank Reversions
This reverts all of the ROMX Bank related changes from the old english translation repo to what they should be based on the main PRET repo. This should help for testing any errors related to space overwriting and whatnot.

New translations exist a few minor places in code comments. But I also translated all of the name screen text so it should now not be the only oddity on those screens.

Translation corrections occurred on a few of the previously translated menu screens. They should now be more in line with the production releases of G1 and G2.